Changed activity point handling - feedback prompt came too early Changed treemap double click handling: MB2 now zooms out, MB3 does nothing

+++ kdirstat/kdirstatsettings.cpp
@@ -4,7 +4,7 @@
  *   License:	GPL - See file COPYING for details.
  *   Author:	Stefan Hundhammer <sh@suse.de>
  *
- *   Updated:	2003-01-07
+ *   Updated:	2003-01-08
  */
 
 
@@ -986,7 +986,7 @@ KTreemapPage::setup()
     _doCushionShading->setChecked	( config->readBoolEntry( "CushionShading"	, true	) );
 
     _ambientLight->setValue		( config->readNumEntry( "AmbientLight"		  , DefaultAmbientLight       ) );
-    _heightScalePercent->setValue( 100 *  config->readDoubleNumEntry ( "HeightScaleFactor", DefaultHeightScaleFactor  ) );
+    _heightScalePercent->setValue( (int) ( 100 *  config->readDoubleNumEntry ( "HeightScaleFactor", DefaultHeightScaleFactor  ) ) );
     _ensureContrast->setChecked		( config->readBoolEntry( "EnsureContrast"	, true	) );
     _forceCushionGrid->setChecked	( config->readBoolEntry( "ForceCushionGrid"	, false ) );
     _minTileSize->setValue		( config->readNumEntry ( "MinTileSize"		, DefaultMinTileSize ) );
